Support unserialization of environments and their contents. (#1)
This handles loading of the environments, their variables and their parent
environments. All unserialized environments are stored in a global Shared
object; the RObject returned by parse_object is just an index into that global
object. This ensures that we faithfully capture REFSXP-based references to the
same environment, and avoids duplication of the environment's variables.

All functions were refactored to accept a Shared object, which keeps track of
the total number of environments for correct REFSXP indexing. This new object
also keeps track of SYMSXP objects, as these also bump the REFSXP counter. 
Like the environments, the RObjects returned by parse_object for symbols are
now just indexes into the Shared object's global symbol cache.

Finally, we dragged in the definition of the pseudo-SEXPTYPEs from the R
source. This eliminates a lot of the empirically determined mystery constants.