Write RDS files from their C++ representations. (#2)
This enables users to construct their own RObject representations in C++, and
then dump it to file via the new write_rds() functions. We update the byteme 
dependency to get the new Gzip writers. 

Moved the Parsed class to its own file (and renamed the class to RdsFile) for
easier import. In particular, we want to avoid writers and readers from
directly including each other's headers, which could get confusing.

Also updated the README with new instructions.