Support parsing and writing of external pointers. (#4)
This is primarily motivated by the experimental hashtab() in utils, which uses
an external pointer as its underlying implementation.  In the general case,
restoring the external pointer itself obviously doesn't do anything, but the
protection value and the tag might tell you something, and that's what the
hashtab uses to store its contents.