Test for correct parsing of objects with wiped attributes.
Specifically, this should have the same behavior as if we never set those
attributes in the first place. This test is motivated by the relatively strange
"memory" of ALTREPs, in that they remember their prior "with attribute" state;
we just want to check that this is not the case for regular objects.