Improve parsing efficiency by switching to byteme::PerByte.
This eliminates the need for the leftover buffer and avoids unnecessary copies
to shift the used bytes in the leftover buffer at every call to the (now
removed) extract_up_to. Such copying caused major perf degradations when
parsing complex objects with lots of nested structure but little actual data. 

In addition, we now have access to PerByteParallel, which allows us to 
read from disk and parse in parallel. This should further improve perf.