Use PerByte::advance_and_extract() to copy byte chunks from the input.
This should be faster than a get()/advance() loop as it uses std::copy() under
the hood. However, it does require upgrading to the latest version of byteme.

Note that we don't do this for strings as we don't have a buffer we can safely
copy the byte chunks into - std::string methods only expose const pointers.